Changes widescreen detection to a store. Tweaks to tooltip delay for apply button.
